GitScriptor-Web/apps/api/pyproject.toml
[tool.poetry]
name = "gitscriptor-api"
version = "0.1.0"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
pydantic = "^2.7"
python-dotenv = "^1.0"
requests = "^2.31.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
authlib = "^1.3.0"
httpx = "^0.27.0"
alembic = "^1.13.0"
cryptography = "^42.0.0"
psutil = "^5.9.0"
gitscriptor_core = {path = "../../packages/gitscriptor_core", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
httpx = "^0.27"
ruff = "^0.4"
black = "^24.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
